Fix protocol checksum byte wraparound with native regression

+++ libfprint-driver/goodix_proto.c
@@ -132,7 +132,7 @@ goodix_decode_protocol (guint8 *data, guint32 data_len, guint8 *cmd,
   *payload = g_memdup2 (data + sizeof (GoodixProtocol), length);
   *payload_len = length;
   *valid_checksum =
-    0xaa - goodix_calc_checksum (data, sizeof (GoodixProtocol) + length) ==
+    (guint8) (0xaa - goodix_calc_checksum (data, sizeof (GoodixProtocol) + length)) ==
     data[sizeof (GoodixProtocol) + length];
   *valid_null_checksum =
     GOODIX_NULL_CHECKSUM == data[sizeof (GoodixProtocol) + length];
